requirements.txt
python-telegram-bot[job-queue]==20.7
gspread==5.11.4
pandas==2.1.4
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
httpx==0.25.2
python-dotenv==1.0.0